X1 clean code
